+++ grafana-plugin/src/pages/users/Users.tsx
@@ -23,7 +23,7 @@ import { User as UserType } from 'models/user/user.types';
 import { PageProps, WithStoreProps } from 'state/types';
 import { withMobXProviderContext } from 'state/withStore';
 import LocationHelper from 'utils/LocationHelper';
-import { isUserActionAllowed, UserActions } from 'utils/authorization';
+import { generateMissingPermissionMessage, isUserActionAllowed, UserActions } from 'utils/authorization';
 import { PLUGIN_ROOT } from 'utils/consts';
 
 import { getUserRowClassNameFn } from './Users.helpers';
@@ -43,6 +43,7 @@ interface UsersState extends PageBaseState {
   usersFilters?: {
     searchTerm: string;
   };
+  initialUsersLoaded: boolean;
 }
 
 @observer
@@ -56,9 +57,10 @@ class Users extends React.Component<UsersProps, UsersState> {
     },
 
     errorData: initErrorDataState(),
+    initialUsersLoaded: false,
   };
 
-  initialUsersLoaded = false;
+  REQUIRED_PERMISSION_TO_VIEW_USERS = UserActions.UserSettingsWrite;
 
   async componentDidMount() {
     const {
@@ -74,18 +76,20 @@ class Users extends React.Component<UsersProps, UsersState> {
     const { usersFilters, page } = this.state;
     const { userStore } = store;
 
-    if (!isUserActionAllowed(UserActions.UserSettingsWrite)) {
+    if (!isUserActionAllowed(this.REQUIRED_PERMISSION_TO_VIEW_USERS)) {
+      this.setState({ initialUsersLoaded: true });
       return;
     }
 
     LocationHelper.update({ p: page }, 'partial');
-    return await userStore.updateItems(usersFilters, page);
+    await userStore.updateItems(usersFilters, page);
+
+    this.setState({ initialUsersLoaded: true });
   };
 
   componentDidUpdate(prevProps: UsersProps) {
-    if (!this.initialUsersLoaded && isUserActionAllowed(UserActions.UserSettingsWrite)) {
+    if (!this.state.initialUsersLoaded) {
       this.updateUsers();
-      this.initialUsersLoaded = true;
     }
 
     if (prevProps.match.params.id !== this.props.match.params.id) {
@@ -117,7 +121,7 @@ class Users extends React.Component<UsersProps, UsersState> {
   };
 
   render() {
-    const { usersFilters, userPkToEdit, page, errorData } = this.state;
+    const { usersFilters, userPkToEdit, page, errorData, initialUsersLoaded } = this.state;
     const {
       store,
       match: {
@@ -165,6 +169,8 @@ class Users extends React.Component<UsersProps, UsersState> {
 
     const { count, results } = userStore.getSearchResult();
 
+    const authorizedToViewUsers = isUserActionAllowed(this.REQUIRED_PERMISSION_TO_VIEW_USERS);
+
     return (
       <PageErrorHandlingWrapper
         errorData={errorData}
@@ -182,10 +188,12 @@ class Users extends React.Component<UsersProps, UsersState> {
                       <LegacyNavHeading>
                         <Text.Title level={3}>Users</Text.Title>
                       </LegacyNavHeading>
-                      <Text type="secondary">
-                        To manage permissions or add users, please visit{' '}
-                        <a href="/org/users">Grafana user management</a>
-                      </Text>
+                      {authorizedToViewUsers && (
+                        <Text type="secondary">
+                          To manage permissions or add users, please visit{' '}
+                          <a href="/org/users">Grafana user management</a>
+                        </Text>
+                      )}
                     </div>
                   </div>
                   <PluginLink query={{ page: 'users', id: 'me' }}>
@@ -194,7 +202,7 @@ class Users extends React.Component<UsersProps, UsersState> {
                     </Button>
                   </PluginLink>
                 </div>
-                {isUserActionAllowed(UserActions.UserSettingsRead) ? (
+                {authorizedToViewUsers ? (
                   <>
                     <div className={cx('user-filters-container')}>
                       <UsersFilters
@@ -213,7 +221,7 @@ class Users extends React.Component<UsersProps, UsersState> {
                     </div>
 
                     <GTable
-                      emptyText={results ? 'No users found' : 'Loading...'}
+                      emptyText={initialUsersLoaded ? 'No users found' : 'Loading...'}
                       rowKey="pk"
                       data={results}
                       columns={columns}
@@ -230,8 +238,9 @@ class Users extends React.Component<UsersProps, UsersState> {
                     /* @ts-ignore */
                     title={
                       <>
-                        You don't have enough permissions to view other users because you are not Admin.{' '}
-                        <PluginLink query={{ page: 'users', id: 'me' }}>Click here</PluginLink> to open your profile
+                        {generateMissingPermissionMessage(this.REQUIRED_PERMISSION_TO_VIEW_USERS)} to be able to view
+                        OnCall users. <PluginLink query={{ page: 'users', id: 'me' }}>Click here</PluginLink> to open
+                        your profile
                       </>
                     }
                     severity="info"