+++ CHANGELOG.md
@@ -34,6 +34,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed duplicate API call, in the UI on plugin initial load, to `GET /api/internal/v1/alert_receive_channels`
 - Increased plugin startup speed ([#1200](https://github.com/grafana/oncall/pull/1200))
 
+### Fixed
+
+- Removed duplicate API call, in the UI on plugin initial load, to `GET /api/internal/v1/alert_receive_channels`
+- Fix minor UI bug on OnCall users page where it would idefinitely show a "Loading..." message
+
 ## v1.1.18 (2023-01-18)
 
 ### Added